+++ packages/cli/CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Multichain support for TypeScript manifest (#2097)
 - Support for multi endpoints CLI deployment (#2117)
 
+### Fixed
+- Building TS manifest on Windows (#2118)
+
 ## [4.0.5] - 2023-10-18
 ### Changed
 - Update error handling on deployment commands (#2108)